Created a seperate file to handle encryption infromation so there would be more space on our main script file. Also added a base frame work for out possible login system
