+++ templates/layout.html
@@ -23,7 +23,7 @@
             <a class="nav-item nav-link" href="/about">About</a>
             <a class="nav-item nav-link" href="/login">Login</a>
             <a class="nav-item nav-link" href="/register">Register</a>
-            <a class="nav-item nav-link" href="/user_page">User Page</a>
+            <a class="nav-item nav-link" href= "/user_page">User Page</a>
             <a class="nav-item nav-link" href="/choose_quiz">Quiz Select</a>
           </div>
         </div>
